\name{FastEPRR-package}
\alias{FastEPRR-package}
\alias{FastEPRR}
\docType{package}
\title{
FastEPRR: a Fast Estimator for the Population Recombination Rate
}
\description{
FastEPRR, an R package for estimating the population recombination rate with the assumption of constant recombination rate over the region.  
}
\details{
\tabular{ll}{
Package: \tab FastEPRR\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2019-06-01\cr
License: \tab GPL\cr
}

}
\author{
Ziqian Hao, Feng Gao

Maintainer: Ziqian Hao <haoziqian@picb.ac.cn>, Haipeng Li <lihaipeng@picb.ac.cn>
}
\references{
}

\keyword{ package }


