\name{FastEPRR}
\alias{FastEPRR}
\title{
FastEPRR: a Fast Estimator for the Population Recombination Rate
}
\description{
FastEPRR, the main function for estimating the population recombination rate with the assumption of constant recombination rate over the region. 
}



\references{

}

\author{
Ziqian Hao, Feng Gao
}


