\name{FastEPRR-package}
\alias{FastEPRR-package}
\alias{FastEPRR}
\docType{package}
\title{
FastEPRR: a Fast Estimator for the Population Recombination Rate
}
\description{
FastEPRR, an R package for estimating the population recombination rate with the assumption of constant recombination rate over the region.  
}
\details{
\tabular{ll}{
Package: \tab FastEPRR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-01\cr
License: \tab GPL\cr
}

}
\author{
Feng Gao

Maintainer: Feng Gao <gaofeng@picb.ac.cn>, Haipeng Li <lihaipeng@picb.ac.cn>
}
\references{
K. Lin, A. Futschik, and H. Li(2013) A fast estimate for the population recombination rate based on regression, Genetics, 194, 473-484.
}

\keyword{ package }


