\name{FastEPRR}
\alias{FastEPRR}
\title{
FastEPRR: a Fast Estimator for the Population Recombination Rate
}
\description{
FastEPRR, the main function for estimating the population recombination rate with the assumption of constant recombination rate over the region. 
}



\references{
Lin, K., Futschik, A. and Li, H. (2013) A fast estimate for the population recombination rate based on regression, Genetics, 194, 473-484.
}

\author{
Gaofeng
}


